--- anti-corruption-layer/etc/api-gateway.urm.puml
@@ -1,2 +1,0 @@
-@startuml
-@enduml
