+++ anti-corruption-layer/src/main/java/com/iluwatar/corruption/system/legacy/Store.java
@@ -1,0 +1,6 @@
+package com.iluwatar.corruption.system.legacy;
+
+import com.iluwatar.corruption.system.DataStore;
+
+public class Store extends DataStore<LegacyOrder> {}
+
