+++ anti-corruption-layer/src/main/java/com/iluwatar/corruption/system/modern/ModernShop.java
@@ -2,11 +2,14 @@ package com.iluwatar.corruption.system.modern;
 
 import com.iluwatar.corruption.system.AntiCorruptionLayer;
 import com.iluwatar.corruption.system.ShopException;
-import com.iluwatar.corruption.system.obsolete.Order;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.Optional;
 
+/**
+ * The class represents a modern shop system.
+ * The main purpose of the class is to place orders and find orders.
+ */
 public class ModernShop {
     @Autowired
     private Store store;
@@ -14,6 +17,11 @@ public class ModernShop {
     @Autowired
     private AntiCorruptionLayer acl;
 
+    /**
+     * Places the order in the modern system.
+     * If the order is already present in the legacy system, then the order is placed only if the data is the same.
+     * If the order is not present in the legacy system, then the order is placed in the modern system.
+     */
     public void placeOrder(ModernOrder order) throws ShopException {
 
         String id = order.getId();
@@ -21,7 +29,7 @@ public class ModernShop {
             throw ShopException.throwDupEx(id);
         }
 
-        Optional<ModernOrder> orderInObsoleteSystem = acl.findOrderInObsoleteSystem(id);
+        Optional<ModernOrder> orderInObsoleteSystem = acl.findOrderInLegacySystem(id);
 
         if (orderInObsoleteSystem.isPresent()) {
             var modernOrder = orderInObsoleteSystem.get();
@@ -35,6 +43,9 @@ public class ModernShop {
         }
     }
 
+    /**
+     * Finds the order in the modern system.
+     */
     public Optional<ModernOrder> findOrder(String orderId) {
         return store.get(orderId);
     }
