+++ anti-corruption-layer/src/main/java/com/iluwatar/corruption/system/modern/Store.java
@@ -1,7 +1,6 @@
 package com.iluwatar.corruption.system.modern;
 
 import com.iluwatar.corruption.system.DataStore;
-import com.iluwatar.corruption.system.obsolete.Order;
 
 
 public class Store extends DataStore<ModernOrder> {}
