--- anti-corruption-layer/src/main/java/com/iluwatar/corruption/system/obsolete/Store.java
@@ -1,7 +1,0 @@
-package com.iluwatar.corruption.system.obsolete;
-
-import com.iluwatar.corruption.system.DataStore;
-
-
-public class Store extends DataStore<Order> {}
-
