+++ api-gateway/api-gateway-service/src/main/java/com/iluwatar/api/gateway/ImageClient.java
@@ -22,26 +22,11 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.iluwatar.corruption;
-
-import lombok.Getter;
-import lombok.Setter;
+package com.iluwatar.api.gateway;
 
 /**
- * Encapsulates all of the information that a desktop client needs to display a product.
+ * An interface used to communicate with the Image microservice.
  */
-@Getter
-@Setter
-public class DesktopProduct {
-
-  /**
-   * The price of the product.
-   */
-  private String price;
-
-  /**
-   * The path to the image of the product.
-   */
-  private String imagePath;
-
+public interface ImageClient {
+  String getImagePath();
 }
