+++ api-gateway/api-gateway-service/src/main/java/com/iluwatar/api/gateway/PriceClientImpl.java
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.iluwatar.corruption;
+package com.iluwatar.api.gateway;
 
 import java.io.IOException;
 import java.net.URI;
@@ -33,35 +33,36 @@ import java.net.http.HttpResponse.BodyHandlers;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
+
 /**
- * An adapter to communicate with the Image microservice.
+ * An adapter to communicate with the Price microservice.
  */
 @Slf4j
 @Component
-public class ImageClientImpl implements ImageClient {
+public class PriceClientImpl implements PriceClient {
 
   /**
-   * Makes a simple HTTP Get request to the Image microservice.
+   * Makes a simple HTTP Get request to the Price microservice.
    *
-   * @return The path to the image
+   * @return The price of the product
    */
   @Override
-  public String getImagePath() {
+  public String getPrice() {
     var httpClient = HttpClient.newHttpClient();
     var httpGet = HttpRequest.newBuilder()
         .GET()
-        .uri(URI.create("http://localhost:50005/image-path"))
+        .uri(URI.create("http://localhost:50006/price"))
         .build();
 
     try {
-      LOGGER.info("Sending request to fetch image path");
+      LOGGER.info("Sending request to fetch price info");
       var httpResponse = httpClient.send(httpGet, BodyHandlers.ofString());
       logResponse(httpResponse);
       return httpResponse.body();
-    } catch (IOException ioe) {
-      LOGGER.error("Failure occurred while getting image path", ioe);
-    } catch (InterruptedException ie) {
-      LOGGER.error("Failure occurred while getting image path", ie);
+    } catch (IOException e) {
+      LOGGER.error("Failure occurred while getting price info", e);
+    } catch (InterruptedException e) {
+      LOGGER.error("Failure occurred while getting price info", e);
       Thread.currentThread().interrupt();
     }
 
@@ -70,9 +71,9 @@ public class ImageClientImpl implements ImageClient {
 
   private void logResponse(HttpResponse<String> httpResponse) {
     if (isSuccessResponse(httpResponse.statusCode())) {
-      LOGGER.info("Image path received successfully");
+      LOGGER.info("Price info received successfully");
     } else {
-      LOGGER.warn("Image path request failed");
+      LOGGER.warn("Price info request failed");
     }
   }
